+++ crates/op-test-vectors/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "op-test-vectors"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+op-alloy-rpc-types = { git = "https://github.com/alloy-rs/op-alloy.git", rev = "97cca5624dc3db8631ff7866e091876fd2fedd25" }
+op-alloy-consensus = { git = "https://github.com/alloy-rs/op-alloy.git", rev = "97cca5624dc3db8631ff7866e091876fd2fedd25" }
+alloy = { version = "0.1", features = ["rpc", "rpc-types-trace"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.117"
+
+[dev-dependencies]
+eyre = "0.6.12"
